правки по комментариям наставника
