+++ src/mock/comments.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { getRandomInteger } from './utils.js';
+import { getRandomInteger } from '../utils.js';
 
 const generateText = () => {
   const texts = [
