+++ src/mock/film.js
@@ -1,5 +1,6 @@
 import { generateComments } from './comments.js';
-import { FILM_CARDS_COUNT, getRandomFloat, getRandomInteger } from './utils.js';
+import {getRandomFloat, getRandomInteger } from '../utils.js';
+import dayjs from 'dayjs';
 
 const GENRES = [
   'action',
@@ -68,9 +69,87 @@ const generateDescription = () => {
   return [...new Set(descriptionArray)].join('');
 };
 
+const generateDirector = () => {
+  const directors = [
+    'Steven Spielberg',
+    'Martin Scorsese',
+    'Quentin Tarantino',
+    'Christopher Nolan',
+    'David Fincher',
+    'James Cameron',
+    'Tim Burton',
+  ];
+  const randomIndex = getRandomInteger(0, directors.length - 1);
+
+  return directors[randomIndex];
+};
+
+const generateWriters = () => {
+  const writers = [
+    'Charlie Kaufman, Spike Jonze, Jane Campion, Billy Wilder, Wes Anderson',
+    'Richard Linklater, Anne Wigton, Heinz Herald, Richard Weil',
+    'Anne Wigton, Spike Jonze, Charlie Kaufman',
+  ];
+
+  const randomIndex = getRandomInteger(0, writers.length - 1);
+
+  return writers[randomIndex];
+};
+
+const generateActors = () => {
+  const actors = [
+    'Adam Sandler, Dwayne Johnson, Ryan Reynolds, Ben Affleck',
+    'Will Smith, Adam Driver, Finn Wolfhard',
+    'Scarlett Johansson, Matt Damon, Millie Bobby Brown',
+    'Reese Witherspoon, Bob Odenkirk, Robert Downey, Jr., James Gunn, Regé-Jean Page, Jennifer Aniston',
+  ];
+
+  const randomIndex = getRandomInteger(0, actors.length - 1);
+
+  return actors[randomIndex];
+
+};
+const generateCountry = () => {
+  const countries = [
+    'Germany',
+    'USA',
+    'Russia',
+    'Italy',
+    'Belgium',
+    'Sweden',
+    'Spain',
+    'China',
+    'Japan',
+    'Ireland',
+  ];
+  const randomIndex = getRandomInteger(0, countries.length - 1);
+
+  return countries[randomIndex];
+};
+
+const generateAgeRating = () => {
+  const ratings = [
+    '0+',
+    '6+',
+    '12+',
+    '16+',
+    '18+',
+  ];
+  const randomIndex = getRandomInteger(0, ratings.length - 1);
+
+  return ratings[randomIndex];
+};
+
+const generateDate = () => {
+  const day = dayjs().date((getRandomInteger(-22265, dayjs().date())));
+  return dayjs(day).format('DD MMMM YYYY');
+};
+
 const generateFilm = () => {
   const runtimeHours = getRandomInteger(1, 3);
   const runtimeMins = getRandomInteger(1, 60);
+  const genresArray = new Array(getRandomInteger(1, GENRES.length)).fill('').map(() =>
+    GENRES[getRandomInteger(0, GENRES.length - 1)]);
   return {
     title: generateTitle(),
     poster: generatePoster(),
@@ -78,47 +157,20 @@ const generateFilm = () => {
     comments: generateComments(),
     rating: getRandomFloat(1, 10, 1),
     releaseYear: getRandomInteger(1960, 2021),
-    runtime: `${runtimeHours}h ${runtimeMins}m`,
-    genre: GENRES[getRandomInteger(0, GENRES.length - 1)],
+    runtime: {runtimeHours, runtimeMins},
+    genres: [...new Set(genresArray)],
+    director: generateDirector(),
+    writers: generateWriters(),
+    actors: generateActors(),
+    release: generateDate(),
+    country: generateCountry(),
+    ageRating: generateAgeRating(),
     isFavorite: Boolean(getRandomInteger(0, 1)),
     isInWatchlist: Boolean(getRandomInteger(0, 1)),
     isWatched: Boolean(getRandomInteger(0, 1)),
   };
 };
 
-const films = new Array(FILM_CARDS_COUNT).fill().map(generateFilm);
+const generateFilms = (filmsCount) => new Array(filmsCount).fill().map(generateFilm);
 
-const getFavorites = () => {
-  const favorites = [];
-  for (const film of films) {
-    if (film.isFavorite) {
-      favorites.push(film);
-    }
-  }
-  return favorites;
-};
-const favoritesCount = getFavorites().length;
-
-const getHistory = () => {
-  const watched = [];
-  for (const film of films) {
-    if (film.isWatched) {
-      watched.push(film);
-    }
-  }
-  return watched;
-};
-const watchedCount = getHistory().length;
-
-const getWatchlist = () => {
-  const watchlist = [];
-  for (const film of films) {
-    if (film.isInWatchlist) {
-      watchlist.push(film);
-    }
-  }
-  return watchlist;
-};
-const watchlistCount = getWatchlist().length;
-
-export {generatePoster, generateTitle, films, GENRES, generateDescription, FILM_CARDS_COUNT, favoritesCount, watchedCount, watchlistCount};
+export {generatePoster, generateTitle, GENRES, generateDescription, generateFilms};
