+++ src/view/menu.js
@@ -1,13 +1,14 @@
-import { favoritesCount, watchedCount, watchlistCount } from '../mock/film.js';
-
-const menuListTemplate = () => (
-  `
+const menuListTemplate = (films) => {
+  const getFilterInFavorites = films.filter((film) => film.isFavorite).length;
+  const getFilterHistory = films.filter((film) => film.isWatched).length;
+  const getFilterInWatchlist = films.filter((film) => film.isInWatchlist).length;
+  return `
   <nav class="main-navigation">
     <div class="main-navigation__items">
       <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-      <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">${watchlistCount}</span></a>
-      <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">${watchedCount}</span></a>
-      <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">${favoritesCount}</span></a>
+      <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">${getFilterInWatchlist}</span></a>
+      <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">${getFilterHistory}</span></a>
+      <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">${getFilterInFavorites}</span></a>
     </div>
       <a href="#stats" class="main-navigation__additional">Stats</a>
   </nav>
@@ -16,7 +17,7 @@ const menuListTemplate = () => (
     <li><a href="#" class="sort__button">Sort by date</a></li>
     <li><a href="#" class="sort__button">Sort by rating</a></li>
   </ul>
- `
-);
+ `;
+};
 
 export {menuListTemplate};
