+++ src/view/stats.js
@@ -1,12 +1,10 @@
-import { watchedCount } from '../mock/film.js';
-
 const statisticsTemplate = () => (
   `
   <section class="statistic">
     <ul class="statistic__text-list">
       <li class="statistic__text-item">
         <h4 class="statistic__item-title">You watched</h4>
-        <p class="statistic__item-text">${watchedCount}<span class="statistic__item-description">movies</span></p>
+        <p class="statistic__item-text"><span class="statistic__item-description">movies</span></p>
       </li>
       <li class="statistic__text-item">
         <h4 class="statistic__item-title">Total duration</h4>
